pyproject.toml
[project]
name = "langgraph-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.1.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
]
